gateway-service/requirements.txt
# 🎬 NarrativeMorph Gateway Service - Requirements
# FastAPI monolitico per hackathon 72h

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI & ML
openai==1.3.5
pillow==10.1.0
requests==2.31.0

# Database (SQLite)
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# Video Processing
opencv-python==4.8.1.78
imageio==2.34.0
moviepy==1.0.3

# Audio Processing
pydub==0.25.1
gTTS==2.4.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Utils
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
